requirements.txt
Flask==3.1.2
python-dotenv==1.1.1
pydantic==2.11.7
gunicorn==23.0.0
sentence-transformers==5.1.0
numpy==2.3.2
chromadb==1.0.20
scikit-learn==1.7.1
langchain==0.1.0
faiss-cpu==1.7.4
openai==1.3.0
transformers==4.30.0
torch==2.0.0
tiktoken==0.5.0
redis==5.0.0
celery==5.3.0

